pyproject.toml
[project]
name = "dardcollect"
version = "0.1.0"
description = "Tools for building a labelled audiovisual dataset from public-domain archive film."
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "internetarchive>=3.0.0",
    "PyYAML",
    "opencv-python",
    "pyyaml>=6.0.1",
    "openai-whisper",
    "scipy",
    "requests",
    "tqdm",
    "onnxruntime-gpu>=1.18.0; sys_platform == 'linux'",
    "onnxruntime>=1.18.0; sys_platform == 'win32'",
    "nvidia-cuda-runtime-cu12==12.4.127; sys_platform == 'linux'",
    "nvidia-cublas-cu12==12.4.5.8; sys_platform == 'linux'",
    "nvidia-cudnn-cu12==9.1.0.70; sys_platform == 'linux'",
    "nvidia-cufft-cu12==11.2.1.3; sys_platform == 'linux'",
    "nvidia-curand-cu12==10.3.5.147; sys_platform == 'linux'",
    "tensorrt==10.1.0; sys_platform == 'linux'",
    "opencv-python-headless>=4.9.0",
    "moviepy",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "torchaudio==2.5.1",

]

[project.optional-dependencies]
dev = [
    "ruff",
    "ty",
]
# GPU acceleration for detection
gpu = [
    "onnxruntime-gpu",
]
# Full tracking (requires compilation)
tracking = [
    "lap",
    "cython_bbox @ git+https://github.com/samson-wang/cython_bbox",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["scripts", "persondet"]

[project.scripts]
extract-person-clips = "scripts.extract_person_clips:main"

[tool.uv]
index-url = "https://download.pytorch.org/whl/cu121"
extra-index-url = ["https://pypi.org/simple"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes (unused imports, undefined names, etc.)
    "I",    # isort (import ordering)
    "UP",   # pyupgrade (modernise Python syntax)
    "RUF",  # ruff-specific rules
    "RET",  # return statement simplifications
]
ignore = [
    "E402",   # module-level import not at top — intentional after gpu_setup
    "RUF002", # ambiguous Unicode in docstrings — intentional math/prose symbols
    "RUF003", # ambiguous Unicode in comments — intentional math/prose symbols
]

[tool.ruff.lint.isort]
known-first-party = ["persondet"]

[tool.ty.environment]
python = ".venv"
